# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/login-logs": {
      "post": {
        "operationId": "LoginLogsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoginLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "login-logs"
        ]
      },
      "get": {
        "operationId": "LoginLogsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "login-logs"
        ]
      }
    },
    "/login-logs/{id}": {
      "get": {
        "operationId": "LoginLogsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "login-logs"
        ]
      },
      "patch": {
        "operationId": "LoginLogsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoginLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "login-logs"
        ]
      },
      "delete": {
        "operationId": "LoginLogsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "login-logs"
        ]
      }
    },
...
    "/user-resets": {
      "post": {
        "operationId": "UserResetsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "user-resets"
        ]
      },
      "get": {
        "operationId": "UserResetsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user-resets"
        ]
      }
    },
    "/user-resets/{id}": {
      "get": {
        "operationId": "UserResetsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user-resets"
        ]
      },
      "patch": {
        "operationId": "UserResetsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user-resets"
        ]
      },
      "delete": {
        "operationId": "UserResetsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user-resets"
        ]
      }
    },
...
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
    "/subcategories": {
      "post": {
        "operationId": "SubCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "subcategories"
        ]
      },
      "get": {
        "operationId": "SubCategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "subcategories"
        ]
      }
    },
    "/subcategories/{id}": {
      "get": {
        "operationId": "SubCategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "subcategories"
        ]
      },
      "patch": {
        "operationId": "SubCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "subcategories"
        ]
      },
      "delete": {
        "operationId": "SubCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "subcategories"
        ]
      }
    },
...
    "/winning-pattern/{id}": {
      "get": {
        "operationId": "WinningPatternController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      },
      "patch": {
        "operationId": "WinningPatternController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWinningPatternDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      },
      "delete": {
        "operationId": "WinningPatternController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern"
        ]
      }
    },
    "/winning-pattern-types": {
      "post": {
        "operationId": "WinningPatternTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWinningPatternTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern-types"
        ]
      },
      "get": {
        "operationId": "WinningPatternTypesController_findAll",
        "parameters": [
          {
            "name": "winning_pattern_id",
            "required": false,
            "in": "query",
            "description": "Filter by winning pattern ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern-types"
        ]
      }
    },
    "/winning-pattern-types/{id}": {
      "get": {
        "operationId": "WinningPatternTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern-types"
        ]
      },
      "patch": {
        "operationId": "WinningPatternTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWinningPatternTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern-types"
        ]
      },
      "delete": {
        "operationId": "WinningPatternTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "winning-pattern-types"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Unique email address of the user",
            "example": "john.doe@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username of the user",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "description": "User password (plaintext here, should be hashed before saving)",
            "example": "MyStrongP@ssw0rd"
          },
          "roles": {
            "description": "List of role IDs or names assigned to the user",
            "example": [
              "admin",
              "editor"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user (unique when provided; required in admin)",
            "example": "+1-555-1234567"
          },
          "profilePhoto": {
            "type": "string",
            "description": "URL of the user profile photo",
            "example": "https://example.com/avatar.jpg"
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "fullName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Unique email address of the user",
            "example": "john.doe@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username of the user",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "description": "User password (plaintext here, should be hashed before saving)",
            "example": "MyStrongP@ssw0rd"
          },
          "roles": {
            "description": "List of role IDs or names assigned to the user",
            "example": [
              "admin",
              "editor"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user",
            "example": "+1-555-1234567"
          },
          "profilePhoto": {
            "type": "string",
            "description": "URL of the user profile photo",
            "example": "https://example.com/avatar.jpg"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Whether the email is verified",
            "example": true
          },
          "phoneVerified": {
            "type": "boolean",
            "description": "Whether the phone is verified",
            "example": false
          },
          "is_deleted": {
            "type": "boolean",
            "description": "Soft delete flag",
            "example": false
          },
          "status": {
            "type": "string",
            "description": "User status",
            "example": "active"
          }
        }
      },
...
      "CreateLoginLogDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User identifier (if user exists)",
            "example": "665f1a4b2c3d4e5f6a7b8c9d"
          },
          "emailAttempted": {
            "type": "string",
            "description": "Email address that was used to attempt login",
            "example": "user@example.com"
          },
          "userAgent": {
            "type": "string",
            "description": "User agent string of the client",
            "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64)..."
          },
          "ipAddress": {
            "type": "string",
            "description": "IP address of the client",
            "example": "192.168.1.100"
          },
          "status": {
            "type": "string",
            "description": "Login attempt status (success, failed, blocked)",
            "example": "success"
          },
          "failureReason": {
            "type": "string",
            "description": "Failure reason if status is not success",
            "example": "Invalid password"
          },
          "deviceName": {
            "type": "string",
            "description": "Device name",
            "example": "iPhone 13 Pro"
          },
          "deviceType": {
            "type": "string",
            "description": "Device type (Mobile, Tablet, Desktop)",
            "example": "Mobile"
          },
          "deviceInfo": {
            "type": "string",
            "description": "Device information (Browser on OS)",
            "example": "Chrome on Windows 10/11"
          },
          "location": {
            "type": "string",
            "description": "Location information",
            "example": "New York, USA"
          },
          "loginAt": {
            "format": "date-time",
            "type": "string",
            "description": "Login timestamp",
            "example": "2025-01-01T10:00:00.000Z"
          }
        },
        "required": [
          "emailAttempted",
          "status"
        ]
      },
      "UpdateLoginLogDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User identifier (if user exists)",
            "example": "665f1a4b2c3d4e5f6a7b8c9d"
          },
          "emailAttempted": {
            "type": "string",
            "description": "Email address that was used to attempt login",
            "example": "user@example.com"
          },
          "userAgent": {
            "type": "string",
            "description": "User agent string of the client",
            "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64)..."
          },
          "ipAddress": {
            "type": "string",
            "description": "IP address of the client",
            "example": "192.168.1.100"
          },
          "status": {
            "type": "string",
            "description": "Login attempt status",
            "example": "success"
          },
          "failureReason": {
            "type": "string",
            "description": "Failure reason if status is not success",
            "example": "Invalid password"
          },
          "deviceName": {
            "type": "string",
            "description": "Device name",
            "example": "iPhone 13 Pro"
          },
          "deviceType": {
            "type": "string",
            "description": "Device type (Mobile, Tablet, Desktop)",
            "example": "Mobile"
          },
          "deviceInfo": {
            "type": "string",
            "description": "Device information (Browser on OS)",
            "example": "Chrome on Windows 10/11"
          },
          "location": {
            "type": "string",
            "description": "Location information",
            "example": "New York, USA"
          },
          "loginAt": {
            "format": "date-time",
            "type": "string",
            "description": "Login timestamp",
            "example": "2025-01-01T10:00:00.000Z"
          }
        }
      },
...
      "CreateUserResetDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User identifier (e.g. User document ID)",
            "example": "665f1a4b2c3d4e5f6a7b8c9d"
          },
          "code": {
            "type": "string",
            "description": "Reset code or token",
            "example": "123456"
          },
          "type": {
            "type": "string",
            "description": "Type of reset (e.g. password, email-verification)",
            "example": "password"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "When this reset link/code expires",
            "example": "2025-12-31T23:59:59.000Z"
          },
          "link": {
            "type": "string",
            "description": "Link sent to the user to perform the reset",
            "example": "https://example.com/reset?token=abc123"
          }
        },
        "required": [
          "user_id",
          "code",
          "type",
          "expiresAt",
          "link"
        ]
      },
      "UpdateUserResetDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User identifier (e.g. User document ID)",
            "example": "665f1a4b2c3d4e5f6a7b8c9d"
          },
          "code": {
            "type": "string",
            "description": "Reset code or token",
            "example": "123456"
          },
          "type": {
            "type": "string",
            "description": "Type of reset (e.g. password, email-verification)",
            "example": "password"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "When this reset link/code expires",
            "example": "2025-12-31T23:59:59.000Z"
          },
          "link": {
            "type": "string",
            "description": "Link sent to the user to perform the reset",
            "example": "https://example.com/reset?token=abc123"
          }
        }
      },
...
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the category",
            "example": "Lodging"
          },
          "description": {
            "type": "string",
            "description": "Short description of the category",
            "example": "Accommodation and lodging facilities"
          },
          "status": {
            "type": "string",
            "description": "Status of the category",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "visibility_type": {
            "type": "string",
            "description": "Visibility type of the category",
            "example": "FREE",
            "enum": [
              "FREE",
              "PREMIUM",
              "ORG_ONLY"
            ]
          }
        }
      },
      "CreateSubCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the sub category",
            "example": "Lodging"
          },
          "description": {
            "type": "string",
            "description": "Short description of the sub category",
            "example": "Accommodation and lodging facilities"
          },
          "status": {
            "type": "string",
            "description": "Status of the sub category",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "category_id": {
            "type": "string",
            "description": "Category ID from categories collection",
            "example": "65f1c2a4b1234567890abcd1"
          }
        },
        "required": [
          "name",
          "category_id"
        ]
      },
      "UpdateSubCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the sub category",
            "example": "Lodging"
          },
          "description": {
            "type": "string",
            "description": "Short description of the sub category",
            "example": "Accommodation and lodging facilities"
          },
          "status": {
            "type": "string",
            "description": "Status of the sub category",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "category_id": {
            "type": "string",
            "description": "Category ID from categories collection",
            "example": "65f1c2a4b1234567890abcd1"
          }
        }
      },
...
      "UpdateWinningPatternDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Line"
          },
          "slug": {
            "type": "string",
            "description": "Slug",
            "example": "line"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "pattern_type": {
            "type": "string",
            "description": "Pattern type: BASIC or ADVANCED",
            "enum": [
              "BASIC",
              "ADVANCED"
            ]
          }
        }
      },
...
    }
  }
}